agent-host/pyproject.toml
[tool.poetry]
name = "agent-host"
version = "0.1.0"
description = "Agent Host - BFF service providing chat interface for MCP Tools Provider"
authors = ["Bruno van de Werve <bvandewe@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
neuroglia-python = { version = "^0.7.5", extras = ["mongodb"] }
redis = "^7.1.0"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
pyjwt = { version = "^2.8.0", extras = ["crypto"] }
httpx = "^0.27.0"
multipledispatch = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
ruff = ">=0.1.0"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 200
select = ["E", "F", "W", "I", "UP"]

[tool.black]
line-length = 200
